Create new ActivityMarker is working
java: 
- Update InscriptionActivityRegister to remove unused  attributres/methods.
- Update InscriptionActivityResource to remove unused imports and services. update addInscriptionToUser method in order to make it work and to return newly created ActivityMarker/InscriptionActivity object

+++ src/main/java/fr/istic/m2/taa/pinit/web/rest/InscriptionActivityResource.java
@@ -7,10 +7,7 @@ import fr.istic.m2.taa.pinit.domain.User;
 import fr.istic.m2.taa.pinit.repository.ActivityRepository;
 import fr.istic.m2.taa.pinit.repository.InscriptionActivityRepository;
 import fr.istic.m2.taa.pinit.repository.UserRepository;
-import fr.istic.m2.taa.pinit.service.ActivityService;
-import fr.istic.m2.taa.pinit.service.InscriptionActivityService;
 import fr.istic.m2.taa.pinit.service.SecurityUtilsService;
-import fr.istic.m2.taa.pinit.service.UserService;
 import fr.istic.m2.taa.pinit.web.rest.exception.BadActivityId;
 import fr.istic.m2.taa.pinit.web.rest.exception.BadInscriptionActivityId;
 import fr.istic.m2.taa.pinit.web.rest.exception.BadUserId;
@@ -35,24 +32,21 @@ public class InscriptionActivityResource {
     private final Logger log = LoggerFactory.getLogger(InscriptionActivityResource.class);
 
     private InscriptionActivityRepository inscriptionActivityRepository;
-    private InscriptionActivityService inscriptionActivityService;
 
     private ActivityRepository activityRepository;
-    private ActivityService activityService;
 
     private UserRepository userRepository;
-    private UserService userService;
 
     private SecurityUtilsService securityUtilsService;
 
 
-    public InscriptionActivityResource(InscriptionActivityRepository inscriptionActivityRepository, InscriptionActivityService inscriptionActivityService, ActivityRepository activityRepository, ActivityService activityService, UserRepository userRepository, UserService userService, SecurityUtilsService securityUtilsService) {
+    public InscriptionActivityResource(InscriptionActivityRepository inscriptionActivityRepository, ActivityRepository activityRepository, UserRepository userRepository, SecurityUtilsService securityUtilsService) {
         this.inscriptionActivityRepository = inscriptionActivityRepository;
-        this.inscriptionActivityService = inscriptionActivityService;
+
         this.activityRepository = activityRepository;
-        this.activityService = activityService;
+
         this.userRepository = userRepository;
-        this.userService = userService;
+
         this.securityUtilsService = securityUtilsService;
     }
 
@@ -74,20 +68,23 @@ public class InscriptionActivityResource {
 
     @RequestMapping(value="/users/{userId}/inscriptions", method = RequestMethod.POST)
     @Secured(Authority.USER)
-    public ResponseEntity addInscriptionToUser(@Valid @RequestBody InscriptionActivityRegister ins) throws BadUserId, BadActivityId, NotAuthorized {
-        Optional<User> potentialUser = userRepository.findUserById(ins.getUserId());
+    public InscriptionActivity addInscriptionToUser(@PathVariable("userId") long userId, @Valid @RequestBody InscriptionActivityRegister ins) throws BadUserId, BadActivityId, NotAuthorized {
+        Optional<User> potentialUser = userRepository.findUserById(userId);
 
+        //On vérifie la présence de l'utilisateur en base.
         if (!potentialUser.isPresent()){
-            throw new BadUserId(ins.getUserId());
+            throw new BadUserId(userId);
         }
 
+        //On vérifie que l'utilisateur qui accède à cette resource est légitime.
         long actualUser = securityUtilsService.getCurrentUserLoginId();
-        if (actualUser != ins.getUserId() && !securityUtilsService.isCurrentUserInRole(Authority.ADMIN)){
+        if (actualUser != userId && !securityUtilsService.isCurrentUserInRole(Authority.ADMIN)){
             throw new NotAuthorized("User not authorized to edit inscriptionActivity of another user");
         }
 
         Optional<Activity> potentialActivity = activityRepository.findById(ins.getActivityId());
 
+        // On vérifie que l'activité choisie existe.
         if (!potentialActivity.isPresent()){
             throw new BadActivityId(ins.getActivityId());
         }
@@ -99,7 +96,8 @@ public class InscriptionActivityResource {
 
         inscriptionActivityRepository.save(inscriptionActivity);
 
-        return ResponseEntity.ok().build();
+        log.debug("InscriptionActivity ID: {}", inscriptionActivity.getId());
+        return inscriptionActivity;
     }
 
     @RequestMapping(value="/inscriptions/{inscriptionId}", method = RequestMethod.DELETE)
@@ -111,6 +109,11 @@ public class InscriptionActivityResource {
             throw new BadActivityId(inscriptionId);
         }
 
+        //On vérifie que l'utilisateur qui accède à cette resource est légitime.
+/*        long actualUser = securityUtilsService.getCurrentUserLoginId();
+        if (actualUser != potentialActivity. && !securityUtilsService.isCurrentUserInRole(Authority.ADMIN)){
+            throw new NotAuthorized("User not authorized to edit inscriptionActivity of another user");
+        }*/
 
 
         inscriptionActivityRepository.deleteById(inscriptionId);
